moving layer state to right folder
